requirements.txt
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
funcsigs==1.0.2
functools32==3.2.3.post2
idna==2.6
jsonschema==2.6.0
logging==0.4.9.6
more-itertools==4.1.0
pluggy==0.6.0
py==1.5.3
pytest==3.5.1
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
